Fix compiler warnings for size_t to int conversions and sscanf
- Animation.h: Change getFrameCount() return type to size_t
- EntityManager.h: Change getEntityCount() return type to size_t
- Animation.cpp: Add explicit static_cast<int> for frame index assignment
- main.cpp: Replace sscanf with std::istringstream for safer parsing

+++ backend/include/entities/EntityManager.h
@@ -57,7 +57,7 @@ public:
     void clear();
 
     // 统计信息
-    int getEntityCount() const { return allEntities_.size(); }
+    size_t getEntityCount() const { return allEntities_.size(); }
     int getAliveEntityCount() const;
 
     // 序列化所有实体
